Update Cli showPlayerData method
We now fill a template read from a file to show the actual player's
info. We removed dead code related to the old version of the above
method.

We added a method in Player class to retrieve some info about the last
picked tile of a player. This will be used to fill the above template.

We've solved an overlap problem between text blocks in
concatenateTextBlocks method in Game.

In FileReader we've renamed readLogoFromTextFile since we now use it to
read also the template for the player's information.

We've fixed wrong mapping in Tile: some tiles had wrong number of warms.